requirements.txt
beautifulsoup4==4.12.0
requests==2.31.0
python-telegram-bot==20.3
python-dotenv==1.0.0
pytest==7.4.0
tweepy==4.14.0

python-telegram-bot==20.3